# api/api-docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AgDx API",
    "description": "AgDx API facilitates sharing of crop pest and disease data between platforms and/or by providing links between complementary ICT tools where appropriate.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0.0-beta1"
  },
...
  "servers": [
    {
      "url": "http://localhost:9100",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/v1/pest/update/{id}": {
      "put": {
        "tags": [
          "Pest data"
        ],
        "summary": "Update pest record",
        "operationId": "updatePestData",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PestEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PestEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/crop/update/{id}": {
      "put": {
        "tags": [
          "Crop data"
        ],
        "summary": "Update crop record",
        "operationId": "updateCropData",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/crop/pest/update/{id}": {
      "put": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Update crop pest record",
        "operationId": "updatePestData_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropPestEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropPestEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/crop/disease/update/{id}": {
      "put": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Update crop disease record",
        "operationId": "updatePestData_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropDiseaseEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropDiseaseEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/pest/add": {
      "post": {
        "tags": [
          "Pest data"
        ],
        "summary": "Add new pest",
        "operationId": "addPestData",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PestEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PestEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/crop/pest/add": {
      "post": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Add new pest",
        "operationId": "addPestData_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropPestEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropPestEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/crop/disease/add": {
      "post": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Add new crop pest record",
        "operationId": "addPestData_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropDiseaseEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropDiseaseEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
    "/api/v1/crop/add": {
      "post": {
        "tags": [
          "Crop data"
        ],
        "summary": "Add new crop",
        "operationId": "addCropData",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CropEntity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "api": [

            ]
          }
        ]
      }
    },
...
    "/api/v1/pest/{pestName}/scientific": {
      "get": {
        "tags": [
          "Pest data"
        ],
        "summary": "Get pests using scientific name",
        "operationId": "getPestByScientificName",
        "parameters": [
          {
            "name": "pestName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pest/{pestName}/common": {
      "get": {
        "tags": [
          "Pest data"
        ],
        "summary": "Get pests using common name",
        "operationId": "getPestByCommonName",
        "parameters": [
          {
            "name": "pestName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagePestEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/crop/{cropName}/scientific": {
      "get": {
        "tags": [
          "Crop data"
        ],
        "summary": "Get crops using scientific name",
        "operationId": "getCropByScientificName",
        "parameters": [
          {
            "name": "cropName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/{cropName}/common": {
      "get": {
        "tags": [
          "Crop data"
        ],
        "summary": "Get crops using common name",
        "operationId": "getCropByCommonName",
        "parameters": [
          {
            "name": "cropName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/crop/pest/{pestName}/scientific": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Get details of specific pest using its scientific name",
        "operationId": "getPestByScientificName_1",
        "parameters": [
          {
            "name": "pestName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CropPestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/pest/{pestName}/common": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Get details of specific pest using its common name",
        "operationId": "getPestByCommonName_1",
        "parameters": [
          {
            "name": "pestName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/pest/vector/{vector}": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Get pests by vector",
        "operationId": "getPestByVector",
        "parameters": [
          {
            "name": "vector",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/pest/reporting-date": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Find crop pests by reporting date range",
        "operationId": "getPestsByReportingDate",
        "parameters": [
          {
            "name": "fromDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/pest/record-date": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Find crop disease by recording date range",
        "operationId": "getPestsByRecordingDate",
        "parameters": [
          {
            "name": "fromDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/pest/observe-date": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Find crop pests by observation date range",
        "operationId": "getPestsByObservedDate",
        "parameters": [
          {
            "name": "fromDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "toDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/pest/management-status/{status}": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Get pests by management status",
        "operationId": "getPestByManagementStatus",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Treated",
                "Untreated"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/pest/habitat/{habitat}": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Get pests by habitat",
        "operationId": "getPestByHabitat",
        "parameters": [
          {
            "name": "habitat",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/pest/detection-status/{status}": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Get pests by detection status",
        "operationId": "getPestByDetectionStatus",
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "Undetected",
                "Detected"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/pest/cultivar/{cultivar}": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Find crop disease by cultivar type",
        "operationId": "getPestsByCultivarName",
        "parameters": [
          {
            "name": "cultivar",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/pest/crop-name/{cropName}": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Find crop disease by crop name",
        "operationId": "getPestsByCropName",
        "parameters": [
          {
            "name": "cropName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/pest/category/{category}": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Get pest by category",
        "operationId": "getPestByCategory",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "Virus",
                  "Bacteria",
                  "Nematode",
                  "Plant",
                  "Insect"
                ]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/pest/admin-level/{adminName}/two": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Find crop pest by administration levels",
        "operationId": "getDiseaseByAdminLevelTwo",
        "parameters": [
          {
            "name": "adminName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/pest/admin-level/{adminName}/three": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Find crop pest by administration levels",
        "operationId": "getDiseaseByAdminLevelThree",
        "parameters": [
          {
            "name": "adminName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crop/pest/admin-level/{adminName}/one": {
      "get": {
        "tags": [
          "Crop pest data"
        ],
        "summary": "Find crop pest by administration levels",
        "operationId": "getDiseaseByAdminLevelOne",
        "parameters": [
          {
            "name": "adminName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropPestEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/crop/disease/admin-level/{adminName}/two": {
      "get": {
        "tags": [
          "Crop disease data"
        ],
        "summary": "Find crop diseases by administration levels",
        "operationId": "getDiseaseByAdminLevelTwo_1",
        "parameters": [
          {
            "name": "adminName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCropDiseaseEntity"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PestEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "commonName": {
            "type": "string",
            "description": "Pest common name",
            "example": "Wooly apple aphid"
          },
          "scientificName": {
            "type": "string",
            "description": "Pest scientific name",
            "example": "Eriosoma lanigerum"
          },
          "order": {
            "type": "string",
            "description": "Pest order",
            "example": "Hemiptera"
          },
          "family": {
            "type": "string",
            "description": "Pest family",
            "example": "Aphididae"
          },
          "genus": {
            "type": "string",
            "description": "Pest genus",
            "example": "Eriosoma"
          },
          "category": {
            "type": "string",
            "description": "Pest family",
            "example": "Insect",
            "enum": [
              "Virus",
              "Bacteria",
              "Nematode",
              "Plant",
              "Insect"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "api": {
        "type": "http",
        "in": "header",
        "scheme": "basic"
      }
    }
  }
}